% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bowl_offers}
\alias{bowl_offers}
\title{Purchase offers for bowl}
\format{
A data frame with 70 rows and 2 variables:
\describe{
  \item{message}{character of purchase offer message}
  \item{condition}{binary label indicating if message is warm or tough}
}
}
\source{
Jeong, M., Minson, J., Yeomans, M. & Gino, F. (2019).

"Communicating Warmth in Distributed Negotiations is Surprisingly Ineffective." Study 3.

Study 3. \url{https://osf.io/t7sd6/}
}
\usage{
bowl_offers
}
\description{
A dataset containing the purchase offer message and a
label indicating if the writer was assigned to be warm (1) or tough (0)
}
\keyword{datasets}
