% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textTools.R
\name{cleanpunct}
\alias{cleanpunct}
\title{Cleaning weird encodings}
\usage{
cleanpunct(text)
}
\arguments{
\item{text}{character Vector of strings to clean.}
}
\value{
character Vector of clean strings.
}
\description{
Handles curly quotes, umlauts, etc.
}
\keyword{internal}
