% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textTools.R
\name{cleantext}
\alias{cleantext}
\title{Clean text}
\usage{
cleantext(ex, language = "english", stop.words = TRUE)
}
\arguments{
\item{language}{string. Default "english".}

\item{stop.words}{logical. Default TRUE}
}
\value{
a character vector
}
\keyword{internal}
