% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posTools.R
\name{coreParser}
\alias{coreParser}
\title{Core Parser}
\usage{
coreParser(text, num_mc_cores = parallel::detectCores())
}
\arguments{
\item{text}{a character vector of texts.}

\item{num_mc_cores}{integer Number of cores for parallelization. Default is parallel::detectCores().}
}
\value{
list of
}
\description{
(deprecated) Part-Of-Speech tagging using Stanford CoreNLP.
}
\keyword{internal}
