% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textTools.R
\name{ctxpand}
\alias{ctxpand}
\title{Contraction Expander}
\usage{
ctxpand(text)
}
\arguments{
\item{text}{a character vector of texts.}
}
\value{
a character vector
}
\description{
Expands Contractions
}
\keyword{internal}
