% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictTools.R
\name{dictWrap}
\alias{dictWrap}
\title{Dictionary Wrapper}
\usage{
dictWrap(text, dict = liwc.lists, binary = F, ...)
}
\arguments{
\item{text}{a character vector of texts.}
}
\description{
background function to load
}
\keyword{internal}
