% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature_table}
\alias{feature_table}
\title{Table of Politeness Features}
\format{An object of class \code{data.frame} with 36 rows and 4 columns.}
\usage{
feature_table
}
\description{
This table describes all the text features extracted in this package.
Meant for the vignette.
}
\keyword{datasets}
