% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{foldset}
\alias{foldset}
\title{Fold Assignment for Cross-Validation}
\usage{
foldset(sizer, nfold, balance = NA)
}
\arguments{
\item{sizer}{number of observations in dataset.}

\item{nfold}{number of outer folds needed.}

\item{balance}{Optional vector of a categorical covariate to stratify fold assignment}
}
\value{
vector of fold IDs
}
\description{
background function to load
}
\keyword{internal}
