% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posTools.R
\name{headTokenGrab}
\alias{headTokenGrab}
\title{Head Token Grab}
\usage{
headTokenGrab(x, data)
}
\arguments{
\item{x}{row in SpaCy parse table}

\item{data}{data.frame SpaCy parse table}
}
\value{
word tagged as referent from dependencey
}
\description{
Compile dependency relation for each row of a SpaCy parse table
}
\keyword{internal}
