% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPlot.R
\name{modelPlot}
\alias{modelPlot}
\title{LASSO Coefficient Plot}
\usage{
modelPlot(model1, counts, model2 = NULL, dat = FALSE)
}
\arguments{
\item{model1}{Trained glmnet model}

\item{counts}{Feature counts - either from training data or test data (choose based on application of interest)}

\item{model2}{Trained glmnet model (optional) If you want the Y axis to reflect a second set of coefficients, instead of feature counts.}

\item{dat}{logical If TRUE, then function will return a list with the data.frame used for plotting, as well as the plot itself.}
}
\value{
ggplot object. Layers can be added like any ggplot object
}
\description{
Plots feature counts and coefficients from a trained LASSO model

This plots the coefficients from a trained LASSO model.
}
