% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textTools.R
\name{negator}
\alias{negator}
\title{Negation Scoping}
\usage{
negator(c.words)
}
\arguments{
\item{text}{character Vector of strings}
}
\value{
list of two character vector strings, based on what's been negated
}
\description{
Figures out which words are negated or not
}
\keyword{internal}
