% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textTools.R
\name{noLeadDash}
\alias{noLeadDash}
\title{Cleaning leading punctuation}
\usage{
noLeadDash(text)
}
\arguments{
\item{text}{character Vector of strings to clean.}
}
\value{
character Vector of clean strings.
}
\description{
Handles interruption dashes.
}
\keyword{internal}
