% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polite_dicts}
\alias{polite_dicts}
\title{Feature Dictionaries}
\format{A list of six \code{quanteda::dictionary} objects}
\usage{
polite_dicts
}
\description{
Six dictionary-like features for the detector: Negations; Pauses; Swearing; Pronouns; Formal Titles; and Informal Titles.
}
\keyword{datasets}
