% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polite_train}
\alias{polite_train}
\title{Pre-Trained Politeness}
\format{
list of two objects. x contains pre-calculated politeness features for each document. y contains standardized human annotations for politeness.
}
\source{
Danescu-Niculescu-Mizil, C., Sudhof, M., Jurafsky, D., Leskovec, J. & Potts, C. (2013). A computational approach to politeness with application to social factors. Proc. 51st ACL, 250-259.
}
\usage{
polite_train
}
\description{
A dataset to train a model for detecting politeness.
}
\keyword{datasets}
