% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politenessPlot.R
\name{politenessPlot}
\alias{politenessPlot}
\title{Politeness plot}
\usage{
politenessPlot(df_polite, split = NULL, split_levels = NULL,
  split_name = NULL, split_cols = c("navy", "firebrick"), top_title = "",
  drop_blank = 0.05)
}
\arguments{
\item{df_polite}{a data.frame with politeness features calculated from a document set, as outputed by \code{\link{politeness}}.}

\item{split}{a vector with exactly two unique values. must have a length equal to the number of documents included in \code{df_polite}.}

\item{split_levels}{character vector of length 2 default NULL. Labels for covariate levels for legend. If NULL, this will be inferred from \code{split}.}

\item{split_name}{character default NULL. Name of the covariate for legend.}

\item{split_cols}{character vector of length 2. Name of colors to use.}

\item{top_title}{character default "". Title of plot.}

\item{drop_blank}{numeric of threshold,}
}
\value{
a ggplot of the magnitude of politeness features split by \code{split}.
}
\description{
Plots the prevalence of politeness features in documents, divided by a binary covariate.
}
\details{
Length of \code{split} must be the same as number of rows of \code{df_polite}.
}
\examples{

}
