% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{receptive_model}
\alias{receptive_model}
\title{Pre-Trained Receptivness Model}
\format{A fitted glmnet model}
\source{
Yeomans, M., Minson, J., Collins, H., Chen, F. & Gino, F. (working paper).

"Conversational Receptiveness: Improving Engagement with Opposing Views"

Study 1. \url{https://osf.io/2n59b/}
}
\usage{
receptive_model
}
\description{
A pre-trained model for detecting conversational receptiveness.
Estimated with glmnet using annotated data from an previous paper.
Primarily for use within the receptiveness() function.
}
\keyword{datasets}
