% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{receptive_names}
\alias{receptive_names}
\title{This is the list of variables to be extracted for the receptiveness algorithm
For internal use only, within the receptiveness() function.}
\format{
Character vector containing variable names
}
\source{
Minson, J., Yeomans, M., Collins, H. & Dorison, C.

"Conversational Receptiveness: Improving Engagement with Opposing Views"
}
\usage{
receptive_names
}
\description{
This is the list of variables to be extracted for the receptiveness algorithm
For internal use only, within the receptiveness() function.
}
\keyword{datasets}
