% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{receptive_polite}
\alias{receptive_polite}
\title{Pre-Trained Receptiveness Data}
\format{
Pre-calculated politeness features for the receptive_train dataset
}
\usage{
receptive_polite
}
\description{
A dataset to train a model for detecting conversational receptiveness.
}
\keyword{datasets}
