% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{receptive_train}
\alias{receptive_train}
\title{Pre-Trained Receptiveness Data}
\format{
A data frame with 543 rows and 2 variables:
\describe{
  \item{text}{character written response about policy disagreement}
  \item{receptive}{numeric average of annotator ratings for "receptiveness"}
}
Primarily for use within the receptiveness() function.
}
\source{
Yeomans, M., Minson, J., Collins, H., Chen, F. & Gino, F. (2020).

"Conversational Receptiveness: Improving Engagement with Opposing Views"

Study 1. \url{https://osf.io/2n59b/}
}
\usage{
receptive_train
}
\description{
A dataset to train a model for detecting conversational receptiveness.
}
\keyword{datasets}
