% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{receptTrain}
\alias{receptTrain}
\title{Receptivness Training Data}
\format{A data frame with 543 rows and 2 variables:
\describe{
  \item{message}{character of response}
  \item{condition}{numeric average rating of writer's receptiveness}
}}
\source{
Yeomans, M., Minson, J., Collins, H., Chen, F. & Gino, F. (working paper).

"Conversational Receptiveness: Improving Engagement with Opposing Views"

Study 1. \url{https://osf.io/2n59b/}
}
\usage{
receptTrain
}
\description{
A dataset containing responses to political arguments and
an average rating of their receptiveness to opposing views.
}
\keyword{datasets}
