% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posTools.R
\name{rowToChar}
\alias{rowToChar}
\title{Row To Char}
\usage{
rowToChar(deps)
}
\arguments{
\item{deps}{row from coreNLP dependency table}

\item{data}{a data.frame}
}
\value{
a character
}
\description{
constructs POS-tagged words from coreNLP parse table.
}
\keyword{internal}
