% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{slogodds}
\alias{slogodds}
\title{Variance-Weighted Log Odds}
\usage{
slogodds(x, y)
}
\arguments{
\item{x}{prevalence in one sample}

\item{y}{prevalence in another sample}
}
\value{
variance-weighted log odds ratio of prevalence across samples
}
\description{
background function to load
}
\keyword{internal}
