% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posTools.R
\name{spacyParser}
\alias{spacyParser}
\title{Spacy Parser}
\usage{
spacyParser(txt, num_mc_cores = parallel::detectCores())
}
\arguments{
\item{txt}{a character vector of texts.}

\item{num_mc_cores}{integer Number of cores for parallelization. Default is parallel::detectCores().}
}
\value{
list of compiled POS-tagged items.
}
\description{
Return POS tags from natural language.
}
\keyword{internal}
