% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{speedDateDates}
\alias{speedDateDates}
\title{SpeedDate Meta Data
Metadata from Ranganath, Jurafsky & McFarland (2009)}
\format{a data.frame with person & group IDs, and other meta-data}
\usage{
speedDateDates
}
\description{
SpeedDate Meta Data
Metadata from Ranganath, Jurafsky & McFarland (2009)
}
\keyword{datasets}
