% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{speedDateTurns}
\alias{speedDateTurns}
\title{SpeedDate Politeness}
\format{a data.frame with person & group IDs, and politeness feature counts}
\usage{
speedDateTurns
}
\description{
Turn-by-turn politeness feature counts (the original text is not shared for privacy reasons).
}
\keyword{datasets}
