% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textTools.R
\name{textcounter}
\alias{textcounter}
\title{Text Counter}
\usage{
textcounter(counted, texts, words = F, fixed = T)
}
\arguments{
\item{words}{logical. Default FALSE.}

\item{fixed}{logical. Default TRUE.}
}
\value{
a numeric vector
}
\keyword{internal}
