% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uk2us}
\alias{uk2us}
\title{UK to US Conversion dictionary}
\format{
A quanteda dictionary with named entries. Names are the US version, and entries are the UK version.
}
\source{
Borrowed from the quanteda.dictionaries package on github (from user kbenoit)
}
\usage{
uk2us
}
\description{
For internal use only. This dataset contains a quanteda dictionary for converting UK words to US words. The models in this package were all trained on US English.
}
\keyword{datasets}
