% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textTools.R
\name{usWords}
\alias{usWords}
\title{UK to US conversion}
\usage{
usWords(text)
}
\arguments{
\item{text}{character Vector of strings to convert to US spelling.}
}
\value{
character Vector of Americanized strings.
}
\description{
background function to load.
}
\keyword{internal}
