% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textTools.R
\name{cleantext}
\alias{cleantext}
\title{Clean Text}
\usage{
cleantext(text, language = "english", stop.words = TRUE)
}
\arguments{
\item{text}{character text to be cleaned}

\item{language}{string. Default "english".}

\item{stop.words}{logical. Default TRUE}
}
\value{
a character vector
}
\description{
Basic text cleaning
}
\keyword{internal}
