% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictTools.R
\name{dictWrap}
\alias{dictWrap}
\title{Dictionary Wrapper}
\usage{
dictWrap(
  text,
  dict = NULL,
  binary = FALSE,
  num_mc_cores = parallel::detectCores(),
  ...
)
}
\arguments{
\item{text}{a character vector of texts.}

\item{dict}{a dictionary class object (see \link[quanteda]{dictionary}) containing dictionaries for six of the politeness features}

\item{binary}{return the prevalence (percent of words) or the presence (yes/no) of a feature in each text?}

\item{...}{arguments passes onto the \code{quanteda:dfm} function}
}
\value{
a matrix with six columns (one for each feature) and a row for every text entered into the function.
}
\description{
background function to load
}
\keyword{internal}
