% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature_table}
\alias{feature_table}
\title{Table of Politeness Features}
\format{A data.frame with information about 36 politeness features.}
\usage{
feature_table
}
\description{
This table describes all the text features extracted in this package. See vignette for details.
}
\keyword{datasets}
