% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPoliteTexts.R
\name{findPoliteTexts}
\alias{findPoliteTexts}
\title{Find polite text}
\usage{
findPoliteTexts(text, covar, ...)
}
\arguments{
\item{text}{a character vector of texts.}

\item{covar}{a vector of politeness labels, or other covariate.}

\item{...}{other arguments passed on to exampleTexts. See exampleTexts for details.}
}
\value{
a ggplot of the prevalence of politeness features, conditional on \code{split}. Features are sorted by variance-weighted log odds ratio.
}
\description{
Deprecated... This function has a new name now. See exampleTexts for details.
}
\examples{
data("phone_offers")
polite.data<-politeness(phone_offers$message, parser="none",drop_blank=FALSE)

findPoliteTexts(phone_offers$message,
                phone_offers$condition,
                type = "most",
                num_docs = 5)


}
