% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phone_offers}
\alias{phone_offers}
\title{Purchase offers for phone}
\format{
A data frame with 355 rows and 2 variables:
\describe{
  \item{message}{character of purchase offer message}
  \item{condition}{binary label indicating if message is warm or tough}
}
}
\source{
Jeong, M., Minson, J., Yeomans, M. & Gino, F. (2019).

"Communicating Warmth in Distributed Negotiations is Surprisingly Ineffective."

Study 1. \url{https://osf.io/t7sd6/}
}
\usage{
phone_offers
}
\description{
A dataset containing the purchase offer message and a
label indicating if the writer was assigned to be warm (1) or tough (0)
}
\keyword{datasets}
