% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politeness.R
\name{politeness}
\alias{politeness}
\title{Politeness Features}
\usage{
politeness(text, parser = c("none", "spacy"), binary = FALSE,
  drop.blank = TRUE)
}
\arguments{
\item{text}{a character vector of texts.}

\item{parser}{a character of which parser to use for text. Default is "none" other options are "core" and "spacy".}

\item{binary}{logical default is FALSE. If TRUE outputed data.frame will only have 0 and 1 as possible values else values will be between 0 and 1.}

\item{drop.blank}{logical dafault is TRUE should columns with only 0 value be removed from outputed data.frame.}
}
\value{
a data.frame of politeness features. Posible columns are
}
\description{
Detects politeness features in text
}
\details{
To use parser = "spacy" requires instalation of
}
\examples{

}
