% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politeness.R
\name{politeness}
\alias{politeness}
\title{Politeness Features}
\usage{
politeness(text, parser = c("none", "spacy"), binary = FALSE,
  drop.blank = TRUE)
}
\arguments{
\item{text}{character vector of texts.}

\item{parser}{character Name of dependency parser to use (see details). Without a dependency parser, some features will be approximated, while others cannot be calculated at all.}

\item{binary}{logical  Return feature counts from each text (default) or a binary indicator for the presence of a feature?}

\item{drop.blank}{logical Should features that were not found in any text be removed from the data.frame? Default is TRUE}
}
\value{
a data.frame of politeness features. Posible columns are listed in LINK_TO_TABLE
}
\description{
Detects politeness features in text
}
\details{
We currently support SpaCy which must be installed separately and prior to running the detector (see example for implementation).
}
\examples{

data("phone_offers")

politeness(phone_offers$message, parser="none",drop.blank=FALSE)

# install.packages("spacyr")
# spacyr::spacy_initialize(python_executable = PYTHON_PATH)
# politeness(phone_offers$message, parser="spacy",drop.blank=FALSE)

}
