% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politenessPlot.R
\name{politenessPlot}
\alias{politenessPlot}
\title{Politeness plot}
\usage{
politenessPlot(polite.data, split = NA, split.levels = NA,
  split.name = NA, split.cols = c("navy", "firebrick"), top.title,
  drop.blank = 0.05)
}
\arguments{
\item{polite.data}{a data.frame with politeness features, as outputed by \code{\link{politeness}}.}

\item{split}{logical vector of how to split \code{polite.data}.}

\item{split.levels}{character vector of length 2. First value correponds to cases where \code{split} is FALSE and second value when TRUE.}

\item{split.cols}{character vector of length 2.}

\item{top.title}{character of title of plot}

\item{drop.bank}{numeric of threshold}
}
\value{
a ggplot plot Showing
}
\description{
Plots
}
\details{
Length of \code{split} must be the same as number of rows of \code{polite.data}.
}
\examples{

}
