% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politenessPlot.R
\name{politenessPlot}
\alias{politenessPlot}
\title{Politeness plot}
\usage{
politenessPlot(df_polite, split = NULL, split_levels = NULL,
  split_name = NULL, split_cols = c("navy", "firebrick"), top_title = "",
  drop_blank = 0.05)
}
\arguments{
\item{df_polite}{a data.frame with politeness features, as outputed by \code{\link{politeness}}.}

\item{split}{a vector with exactly two unique values. of how to split \code{df_polite}.}

\item{split_levels}{character vector of length 2 default NULL. Labels of split values for graph legend. If NULL, this will be infered from \code{split}.}

\item{split_name}{character default NULL. with the name of the split variable for the graph legend.}

\item{split_cols}{character vector of length 2.}

\item{top_title}{character default "". Title of plot.}

\item{drop_blank}{numeric of threshold,}
}
\value{
a ggplot of the magnitude of politeness features split by \code{split}.
}
\description{
Plots magnitude of politeness features given a split of data.
}
\details{
Length of \code{split} must be the same as number of rows of \code{df_polite}.
}
\examples{

}
