% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politenessPlot.R
\name{politenessPlot}
\alias{politenessPlot}
\title{Politeness plot}
\usage{
politenessPlot(df_polite, ...)
}
\arguments{
\item{df_polite}{a data.frame with politeness features calculated from a document set, as output by \code{\link{politeness}}.}

\item{...}{other arguments passed on to featurePlot. See featurePlot for details.}
}
\value{
a ggplot of the prevalence of politeness features, conditional on \code{split}. Features are sorted by variance-weighted log odds ratio.
}
\description{
Deprecated... This function has a new name now. See featurePlot for details.
}
\examples{

data("phone_offers")

polite.data<-politeness(phone_offers$message, parser="none", drop_blank=FALSE)

politeness::politenessPlot(polite.data,
                           split=phone_offers$condition,
                           split_levels = c("Tough","Warm"),
                           split_name = "Condition",
                           top_title = "Average Feature Counts")

}
