% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{receptive_model}
\alias{receptive_model}
\title{A pre-trained model for detecting conversational receptiveness.
Estimated with glmnet using annotated data from a previous paper.
Primarily for use within the receptiveness() function.}
\format{
A fitted glmnet model
}
\source{
Minson, J., Yeomans, M., Collins, H. & Dorison, C.

"Conversational Receptiveness: Improving Engagement with Opposing Views"
}
\usage{
receptive_model
}
\description{
A pre-trained model for detecting conversational receptiveness.
Estimated with glmnet using annotated data from a previous paper.
Primarily for use within the receptiveness() function.
}
\keyword{datasets}
