% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacyParser.R
\name{spacyParser}
\alias{spacyParser}
\title{Spacy Parser}
\usage{
spacyParser(txt)
}
\arguments{
\item{txt}{a character vector of texts.}
}
\value{
list of compiled POS-tagged items.
}
\description{
Return POS tags from natural language.
}
\keyword{internal}
